ref(eventstream): Refactor out `SnubaProtocolEventStream` from HTTP and Kafka transports
